+++ internal/db/pull/pull.go
@@ -104,7 +104,7 @@ func dumpRemoteSchema(p utils.Program, ctx context.Context, path string, config 
 func diffRemoteSchema(p utils.Program, ctx context.Context, schema []string, path string, config pgconn.Config, fsys afero.Fs) error {
 	w := utils.StatusWriter{Program: p}
 	// Diff remote db (source) & shadow db (target) and write it as a new migration.
-	output, err := diff.DiffDatabase(ctx, schema, config, w, fsys)
+	output, err := diff.DiffDatabase(ctx, schema, config, w, fsys, diff.DiffSchemaMigra)
 	if err != nil {
 		return err
 	}
